requirements.txt
amqp==2.3.2
aniso8601==3.0.2
asn1crypto==0.24.0
bcrypt==3.1.4
billiard==3.5.0.4
cached-property==1.5.1
cachetools==2.1.0
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
click==6.7
cryptography==2.3.1
demjson==2.2.4
docker==2.7.0
docker-compose==1.18.0
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.6
eventlet==0.22.1
Faker==0.7.18
Flask==1.0.2
Flask-JWT==0.3.2
Flask-Login==0.4.0
flask-mongoengine==0.9.5
Flask-RESTful==0.3.6
Flask-SocketIO==2.9.6
Flask-Testing==0.6.2
Flask-WTF==0.14.2
google-auth==1.5.1
greenlet==0.4.12
gunicorn==19.6.0
holdup==1.6.0
idna==2.7
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
kombu==4.2.1
kubernetes==5.0.0
lxml==4.2.5
MarkupSafe==1.0
marshmallow==2.14.0
mongoengine==0.13.0
oauthlib==2.1.0
paramiko==2.4.2
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyJWT==1.4.2
pymongo==3.4.0
PyNaCl==1.3.0
pyOpenSSL==18.0.0
python-dateutil==2.7.3
python-engineio==2.2.0
python-socketio==2.0.0
pytz==2018.5
pyvmomi==6.5
PyYAML==3.13
redis==2.10.6
requests==2.13.0
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
rsa==3.4.2
six==1.11.0
texttable==0.9.1
urllib3==1.23
vine==1.1.4
websocket-client==0.53.0
Werkzeug==0.14.1
WTForms==2.2.1
xpinyin==0.5.6
yunpian-python-sdk==1.0.0
